Add GitHub workflow to PancakeRecipe solution and update publish path
A GitHub workflow file has been added to the PancakeRecipe solution. This will enable automatic build and deployment on GitHub. In addition, the path of the dotnet publish command has been updated to point to 'PancakeRecipeApi' instead of 'myapp'. As a result, the output directory has been correctly named to reflect the application being built.